Revert "Changed pointer size to uintptr_t type, which should be architechture independent" This apparently causes some implementations of DivGeo to fail when creating objects, among other things. Why remains a mystery.
This reverts commit bd8071725b25254fffba94f11f2f615e971394e6.

+++ src/dgfile.c
@@ -1,6 +1,5 @@
 #include "dg.h"
 #include "git_version.h"
-#include <stdint.h>
 
 #define FILE_VERSION 115
 #define DG_VERSION_NEEDED 15000
@@ -86,13 +85,13 @@ static struct _NameRec meshSlidingModes[]={
 
 static struct _FlagsRec surfaceZoneFlags[]={
   {SZF_LIMITBYSURFACE,'i'},
-  {0,(uintptr_t) NULL}
+  {0,(int) NULL}
 };
 
 static struct _FlagsRec gridPointSegFlags[]={
   {GPSF_USED,'u'},
   {GPSF_TARGET_CW,'t'},
-  {0,(uintptr_t) NULL}
+  {0,(int) NULL}
 };
 
 App LoadNormalApp(XApp xap,char* fName,char** pMsg,int* err) {
