+++ src/vacreate.c
@@ -63,7 +63,6 @@
 #include <values.h>
 #include <stdlib.h>
 #include <stdio.h>
-#include <stdint.h>
 
 #include "vacreate.h"
 
@@ -458,7 +457,7 @@ printf("creating: %s\n",strName);               /* relcheck_ignore_line */
       case '=':
         pt1=va_arg(*pvl,XtPointer);
         if (XmIsToggleButton(w))
-          XmToggleButtonSetState(w,(uintptr_t)pt1,True);
+          XmToggleButtonSetState(w,(int)pt1,True);
         else if (XmIsText(w))
           XmTextSetString(w,pt1);
         else if (XmIsTextField(w))
@@ -609,7 +608,7 @@ int GetResourceInt(Widget wg,char* name,char* class,int def) {
   if (class==NULL) class=name;
   xtr.resource_name=name;
   xtr.resource_class=class;
-  xtr.default_addr=(XtPointer)(uintptr_t)def;
+  xtr.default_addr=(XtPointer)def;
   XtGetApplicationResources(wg,&i,&xtr,1,NULL,0);
 
   return i;
@@ -661,14 +660,14 @@ void CbFreeGC(Widget wg,XtPointer xtpGC,XtPointer pcbs) {
 void CbRemovePTimeOut(Widget wg,XtPointer xtppTO,XtPointer pcbs) {
   XtIntervalId* pxtiid=(XtIntervalId*)xtppTO;
 
-  if (*pxtiid!=(uintptr_t) NULL) {
+  if (*pxtiid!=(int) NULL) {
     XtRemoveTimeOut(*pxtiid);
-    *pxtiid=(uintptr_t) NULL;
+    *pxtiid=(int) NULL;
   }
 }
 
 void CbExitApp(Widget wg,XtPointer xtpExitCode,XtPointer pcbs) {
-  exit((uintptr_t)xtpExitCode);
+  exit((int)xtpExitCode);
 }
 
 void CbDebugPrint(Widget wg,XtPointer xtpDebugStr,XtPointer pcbs) {
@@ -1118,8 +1117,8 @@ void Form2Table(Widget form) {
   for (i=0;i<numChildren;i++) {
     pos[i].wg=children[i];
     GetValues(children[i],XmNuserData,&xtp,NULL);
-    pos[i].x=(uintptr_t)xtp/256;
-    pos[i].y=(uintptr_t)xtp%256;
+    pos[i].x=(int)xtp/256;
+    pos[i].y=(int)xtp%256;
   }
 
   CreateFormTable(form,pos,numChildren);
