+++ src/x_vars.c
@@ -1,5 +1,4 @@
 #include "x_dg.h"
-#include <stdint.h>
 
 /**********************************************************************\
 *                                                                      *
@@ -485,7 +484,7 @@ static void ResetVarsDlgField(VarsEditField vef,unsigned flags) {
     else {
       SetLabelString(vef->dlg->wMessage,s);
       SetUserData(vef->dlg->wMessage,
-	  (XtPointer)(uintptr_t)(flags & RF_NEWVAL? XT_NEWVAL : XT_VEF));
+          (XtPointer)(flags & RF_NEWVAL? XT_NEWVAL : XT_VEF));
     }
   }
 
@@ -731,7 +730,7 @@ static void CbVarsEditFocus(Widget wg,XtPointer pField,XtPointer pcbs) {
 static void CbVarsEditCompare_p(Widget wg,XtPointer xtpDlg,XtPointer pcbs) {
   VarsEditDlg dlg=(VarsEditDlg)xtpDlg;
   VarsEditField vef=dlg->popupMenuField;
-  int mode=(uintptr_t)GetUserData(wg);
+  int mode=(int)GetUserData(wg);
   int i1,i2,comp;
   double f1,f2;
   char* s,* s1;
